@@ -16,10 +16,13 @@
     "qs": "^5.1.0"
   },
   "peerDependencies": {
-    "react": ">=0.13.3"
+    "react": ">=0.14.7"
   },
   "devDependencies": {
     "browserify": "^11.1.0",
+    "eslint": "^2.2.0",
+    "eslint-config-standard": "^5.1.0",
+    "eslint-plugin-standard": "^1.3.2",
     "watchify": "^3.4.0"
   }
 }
