@@ -1,10 +1,9 @@
-var React = require('react');
-var createElement = React.createElement;
-var createClass = React.createClass;
-var PropTypes = React.PropTypes;
+var React = require('react')
+var createElement = React.createElement
+var createClass = React.createClass
+var PropTypes = React.PropTypes
 
 module.exports = createClass({
-
   displayName: 'Lazy',
 
   contextTypes: {
@@ -12,13 +11,10 @@ module.exports = createClass({
   },
 
   render: function () {
-
     return createElement(
       this.context.childComponent,
       this.props,
       this.props.children
-    );
-
+    )
   }
-
-});
+})
