@@ -1,12 +1,10 @@
 var Empty = require('react')
   .createClass({
-
     displayName: 'Empty',
 
     render: function () {
-      return null;
+      return null
     }
+  })
 
-  });
-
-module.exports = Empty;
+module.exports = Empty
