@@ -1,125 +1,109 @@
-var page = require('page');
-var Route = require('./Route');
-var qs = require('qs');
-var assign = require('object-assign');
-var pathToRegexp = require('path-to-regexp');
-var routes = {};
-var config = {base: ''};
-Route.routes = routes;
+var page = require('page')
+var Route = require('./Route')
+var qs = require('qs')
+var assign = require('object-assign')
+var pathToRegexp = require('path-to-regexp')
+var routes = {}
+var config = {base: ''}
+Route.routes = routes
 
-function dropTrailingSlash(str) {
-  return (str || '').replace(/\/$/, '');
+function dropTrailingSlash (str) {
+  return (str || '').replace(/\/$/, '')
 }
 
-function setConfig(opts) {
-
-  assign(config, opts);
-  config.base = dropTrailingSlash(config.base);
-  page(config);
-
+function setConfig (opts) {
+  assign(config, opts)
+  config.base = dropTrailingSlash(config.base)
+  page(config)
 }
 
-function reactPage() {
-
-  var args = Array.prototype.slice.call(arguments);
+function reactPage () {
+  var args = Array.prototype.slice.call(arguments)
 
   if (args.length === 1 && args[0] && typeof args[0] === 'object') {
-    return setConfig(opts);
+    return setConfig(opts)
   }
 
-  if (args.length < 3) return page.apply(this, args);
+  if (args.length < 3) return page.apply(this, args)
 
-  var uri = args.shift();
-  var name;
+  var uri = args.shift()
+  var name
 
   if (typeof args[0] === 'string') {
+    name = uri
+    uri = args.shift()
 
-    name = uri;
-    uri = args.shift();
-
-    routes[name] = pathToRegexp.compile(uri);
-
+    routes[name] = pathToRegexp.compile(uri)
   }
 
-  var callback = args.pop();
-  var components = args;
+  var callback = args.pop()
+  var components = args
 
   page(uri, function (ctx) {
-
-    ctx.name = name;
-    ctx.query = qs.parse(ctx.querystring);
-    ctx.config = config;
+    ctx.name = name
+    ctx.query = qs.parse(ctx.querystring)
+    ctx.config = config
 
     setTimeout(function () {
-
       callback(
         Route(
           components.concat(),
           ctx
         )
-      );
-
-    }, 0);
-
-  });
-
+      )
+    }, 0)
+  })
 }
 
-function transitionTo(name, params, query) {
-
-  var compiler = routes[name];
-  var queryString = typeof query === 'object' ? qs.stringify(query) : '';
+function transitionTo (name, params, query) {
+  var compiler = routes[name]
+  var queryString = typeof query === 'object' ? qs.stringify(query) : ''
 
   page(
     (compiler ? compiler(params) : name) +
     (queryString ? '?' + queryString : '')
-  );
-
+  )
 }
 
-function navigateTo(name, params, query) {
+function navigateTo (name, params, query) {
+  params = assign({}, params)
+  query = assign({}, query)
 
-  params = assign({}, params);
-  query = assign({}, query);
+  var nav = {}
 
-  var nav = {};
-
-  function setParams(p, x) {
-
+  function setParams (p, x) {
     if (typeof p === 'object') {
-      assign(params, p);
+      assign(params, p)
     } else {
-      params[p] = x;
+      params[p] = x
     }
 
-    return nav;
+    return nav
   }
 
-  function setQuery(q, x) {
-
+  function setQuery (q, x) {
     if (typeof q === 'object') {
-      assign(query, q);
+      assign(query, q)
     } else {
-      params[q] = x;
+      params[q] = x
     }
 
-    return nav;
+    return nav
   }
 
-  nav.with = setParams;
-  nav.where = setQuery;
-  nav.and = nav;
+  nav.with = setParams
+  nav.where = setQuery
+  nav.and = nav
 
   nav.go = function () {
-    transitionTo(name, params, query);
-  };
+    transitionTo(name, params, query)
+  }
 
-  return nav;
-
+  return nav
 }
 
-Route.transitionTo = transitionTo;
-reactPage.transitionTo = transitionTo;
+Route.transitionTo = transitionTo
+reactPage.transitionTo = transitionTo
 
 reactPage.navigate = {
   to: navigateTo,
@@ -127,80 +111,73 @@ reactPage.navigate = {
   from: function (ctx) {
     return {
       to: function (name) {
-        return navigateTo(name, ctx.params, ctx.query);
+        return navigateTo(name, ctx.params, ctx.query)
       }
     }
   }
-};
+}
 
-reactPage.use = {};
+reactPage.use = {}
 
 reactPage.set = function (render) {
+  var components = []
 
-  var components = [];
+  var three = {}
 
-  var three = {};
-
-  three.run = setConfig;
+  three.run = setConfig
 
   three.on = function () {
-
-    var args = Array.prototype.slice.call(arguments);
-    var optList = components.concat();
+    var args = Array.prototype.slice.call(arguments)
+    var optList = components.concat()
 
     if (typeof args[1] === 'string') {
-      optList.unshift(args[1]);
-      args.splice(1, 1);
+      optList.unshift(args[1])
+      args.splice(1, 1)
     }
 
-    optList.unshift(args[0]);
+    optList.unshift(args[0])
 
-    args.shift();
-    args.push(render);
+    args.shift()
+    args.push(render)
 
-    optList = optList.concat(args);
+    optList = optList.concat(args)
 
-    reactPage.apply(null, optList);
+    reactPage.apply(null, optList)
 
-    return three;
-  };
+    return three
+  }
 
   three.rewind = function () {
+    components = []
 
-    components = [];
-
-    return three;
-  };
+    return three
+  }
 
   three.with = function () {
+    components = components.concat(Array.prototype.slice.call(arguments))
 
-    components = components.concat(Array.prototype.slice.call(arguments));
+    return three
+  }
 
-    return three;
+  return three
+}
 
-  };
+reactPage.RouteHandler = require('./RouteHandler')
+reactPage.Link = require('./Link')
+reactPage.when = require('./Async')
 
-  return three;
-};
-
-reactPage.RouteHandler = require('./RouteHandler');
-reactPage.Link = require('./Link');
-reactPage.when = require('./Async');
-
-reactPage.redirect = page.redirect;
-reactPage.show = page.show;
-reactPage.start = setConfig;
-reactPage.stop = page.stop;
+reactPage.redirect = page.redirect
+reactPage.show = page.show
+reactPage.start = setConfig
+reactPage.stop = page.stop
 
 reactPage.base = function (str) {
+  str = dropTrailingSlash(str)
 
-  str = dropTrailingSlash(str);
+  assign(config, {base: str})
+  return page.base(str)
+}
 
-  assign(config, {base: str});
-  return page.base(str);
+reactPage.exit = page.exit
 
-};
-
-reactPage.exit = page.exit;
-
-module.exports = reactPage;
+module.exports = reactPage
